restructure code to use moodules
