+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "liveServer.settings.port": 5502
+}
