+++ index.html
@@ -111,6 +111,6 @@
     <footer class="footer">
       <p class="footer-text">copyright: Abass Olanrewaju Waiu</p>
     </footer>
-    <script src="./js/index.js"></script>
+    <script type="module" src="/index.js"></script>
   </body>
 </html>
